Quick idea for player death drops.
Inspired by XC 2 meant to be Malos' core crystal. players drop core crystals on death which can be used to make core ingots. I wanted to add the enchanted glow effect to the graphic but I don't know how. This is probably too complicated, but I thought it would be cool if core crystals could save information like skin and equiptment of killed players and could be used to summon mob look-a-likes as summoned allies.